requirements.txt
glob
numpy
pandas
scikit-learn
tensorflow-cpu
tqdm
